requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
torch==2.1.1
joblib==1.3.2

# Streaming & Messaging
kafka-python==2.0.2
redis==5.0.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
influxdb-client==1.38.0

# Monitoring
prometheus-client==0.19.0
grafana-client==3.7.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.12.0
mypy==1.7.1
pre-commit==3.6.0
ipython==8.18.1
jupyter==1.0.0